fixing ci issues and renaming file
