+++ runtime/hash_impl.h
